test(custom_openai): cover model discovery branches
